🛠️ refactor: move GitResetManager to editor directory structure
