+++ src/editor/git/reset-manager.ts
